protocolDetector: Remove unneeded poll flags
Protocol detector was using uneeded poll flags in the eventmask; POLLHUP
and POLLERR are output only, as documented in poll(2) and Python library
reference:
https://docs.python.org/2.6/library/select.html#select.poll.register

POLLPRI is needed only to detect out-of-band data, which the other side
can send using MSG_OOB flag, as documented in send(2). Since we don't
use out-of-band data, we don't need to check for it.

Since the READ_ONLY_MASK contains only select.POLLIN, and the
READ_WRITE_MASK was used only once, they were removed.  Developers
maintaing this code are expected to know these constants.

Change-Id: I94c535c7aa94410eb3158b298c3df3d1738abe74
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33912
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>