libvirt: do not always log failed libvirt calls
The libvirtconnection module provides libvirt connection handling
for VDSM. libvirt connection and domain objects have their callable
methods transparently wrapped, to let VDSM automatically
detect and react to disconnection.

A side effect of this wrapping is that exceptions are intercepted
before to be reraised and delivered to the caller, and a
quite cryptic and not very informative message is logged for
each exception.

Given that the wrapper is -and has to be- generic, this logging
has no choice in turn but to be generic, but this adds little
value to logs, and surely adds noise and concern to the user.

Proper action must be done in the client code, which must handle
exception correctly and do whatever it is needed, maybe
just log the error, maybe silencing it.

So, this patch just drops the extra debug log.

Change-Id: I170e1870e1d4759c1c7310d71f00450b6e560690
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33937
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>