nwfilter: Do not attempt to re-define the nwfilter when not undefining
Undefining a filter is an operation that can fail. The fact that we
had the same exception swallowing for two operations, looking up and
discovering led us to the wrong assumption then of defining again the
filter even when it had been looked up fine but could not be
undefined.

The most sensible behavior to take for 3.5 is to just consider the
failure to undefine properly and re-define it only in the case of
failure to look up.

Change-Id: Ia284c0554ef86a4668808b26523fbb3b51baebcc
Bug-Url: https://bugzilla.redhat.com/1150718
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33993
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34206