supervdsmServer: Do not use udev to set permissions on external luns
Before plugging external luns to vms, we setup up a temporary udev rule
for setting device permissions. Using this rule will cause the device to
loose the libvirt selinux label if a device has a change event.

This is the same issue we had with vdsm images, but with external luns
we do not trigger change events, so the issue is unlikely. However, if
it happens, it will cause a vm to pause.

This patch applies the same fix, use chown instead of udev USER and
GROUP directives.

Change-Id: I7adb8a441497c9dd335119f55af73189ed89c23c
Relates-To: http://gerrit.ovirt.org/33875
Relates-To: https://bugzilla.redhat.com/1147910
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33994
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>