setupNetworks: doesn't remove broken nw from runningConfig
This patch extends http://gerrit.ovirt.org/#/c/22354/:
network/api.py:setupNetworks() removes broken networks - if network
is not listed in netinfo, but it's in libvirt, _delBrokenNetwork()
is called.

The problem is that when the network doesn't have an existing device,
we only remove it from libvirt and leave it in RunningConfig.

TODO: this problem occurs on 3.4 with libvirt 1.2.9

Now _delBrokenNetwork() remove broken network from runningConfig.

NOTE: This patch doesn't handle the case, when the network is only in
RunningConfig, not in libvirt, not in netinfo. This should be solved in
a separate patch.

Change-Id: I0c507626705d7ead84db2f3aa15e4032f9558d12
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/33995
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>