VDSM hook for random network allocation during VM power on
This hook is useful in cases where one has a set
of equivalent networks, and wants to distribute
VMs evenly among them, so the networks share the
traffic
load.
The more VMs allocated with networks, the
tendency of VMs to be evenly distributed among
networks rise.
The hook uses CustomDeviceProperties and a property
called 'equivnets' which is a whitespace separated list
of network names. When the VM is started, a random
network is allocated to it.

Change-Id: Ib14c76994918750e7eef460411acc337599b0186
Bug-Url: https://bugzilla.redhat.com/1065405
Signed-off-by: ibarkan <ibarkan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34114
Reviewed-by: Dan Kenigsberg <danken@redhat.com>