core: Add python warnings boilerplate
In order to use python's built-in warnings module there some things you
have to set up so they appear in the log.

The options devel/enable_python_warnings in the config is responsible
for turning the warnings on and is turned-off by defaults.

Also enabled warnings for unit tests by defaults.

Change-Id: Ibfb55d82db35efa14dc36f3f748ec01259109038
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34034
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>