tests: Fix patchDecorators so decorated function will actually run
Commit 5c5a0d9c5 introduced a fancy decorator for cleaning up the mess
when we monkey patch conifguratiors. Unfortunatly, this decorator has an
unexpected side effect, of *not* running the decorated function, so the
test would always pass, even if the code is broken.

This patch replaces this decorator with a much simpler one that let the
real decorator (MonkeyPatch) do the right thing.

Change-Id: I1c4be2a15160be48b56a7df2f406e4e0afabd11e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34153
Reviewed-by: mooli tayer <mtayer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>