multiapth: Disable hba rescanning by default
Since commit 674337d8f9 we are scanning both iscsi bus and hbas when
issuing multipath.rescan(). We did not find any bad effects when testing
in our QE lab, but we have reports from customers that these rescans have
bad effects on FC connections.

This patch disables hba rescanning by default, so users can use this
feature if works for them, until we investigate this issue and provide
better solution.

This should have been part of the original patch, since we cannot test
all kinds of storage servers in the same setup and load used by all
users. If hba rescanning cannot be used with certain storage server, it
should be possible to disable this feature.

Change-Id: I812308e54ec2a9eab1ca4115e0456b2dd3990cfb
Bug-Url: https://bugzilla.redhat.com/1152587
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34176
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34215
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>