libvirtconnection: add function to create a new connection
Added a new function open_connection(), which accept credentials and uri,
This will open a new libvirt connection and will not be handled by the
wrapMethod() and will not be stored by the __connection dictionary.

The new function will be needed by the v2v future patches which will need
to connect to a remote libvirt given an uri.

Signed-off-by: Shahar Havivi <shaharh@redhat.com>
Change-Id: I3481d56ee19afec14b36b98b412278148b26379e
Reviewed-on: http://gerrit.ovirt.org/34293
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Shahar Havivi <shavivi@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>