rpc: Replace misuse of assert with TypeError
Asserts are optimized out when running with -O or when using
PYTHONOPTIMIZE environment variable.  This assert seems like a security
related check, and it should never be right to ignore this check.

Replaced with TypeError, which is the expected error when passing
non-existing argument to a function.

I'm not very happy about this, since this hides a security failure,
pretending to be a Python calling error. Anyway this is better then
the assert.

Change-Id: I8c60c65b283f4343448bb9eaf6ccf2526b188db0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29307
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>