vm: log xml from/to libvirt at 'info' level
The XML which VDSM sends or receive to libvirt
is a critical piece of information for troubleshooting
issues, and we must make sure to log it unless the user
really want VDSM to be really silent.

So, bump the relevant logs from 'debug' to 'info',
even if we know the xml dump is actually debug
information.

Change-Id: Ia07823512a8cdf5f05789b29bf3d4daa1a8ccb07
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34358
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>