openstacknet: Fix migration when using security groups
When migrating, the destination libvirt receives an xml already
altered by before_device_create like this test one:

    <?xml version="1.0" encoding="utf-8"?>
    <interface type="bridge">
        <mac address="00:1a:4a:16:01:51"/>
        <model type="virtio"/>
        <source bridge="qbrtest_port_i"/>
        <target dev="taptest_port_i"/>
    </interface>

The issue is that before_device_create hooking point is not part
of the migration and it is the only thing that creates the security
groups bridge and the necessary veths. In order to have migration
working, then, it was necessary to add a hook that on the hooking
point 'before_device_migrate' does the security groups bridge and
veths creation.

Change-Id: Icd8a789c4565f32b32965af3966a4edd361949ea
Bug-Url: https://bugzilla.redhat.com/1048880
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34406
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35961
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>