virt: move all_channels into DomainDescriptor
before the introduction of the DomainDescriptor
class, we had a bunch of functions/methods who
repeteadly parsed the VM XML Domain using the slow
minidom module.

The introduction of the DomainDescriptor opens
the gate for both more efficient and cleaner processing.

DomainDescriptor should provide methods to return
high-level objects, not low-level ones, e.g, XML document
nodes, which are inevitabily tied to the underlying xml
library being used.

The first step in this (long) journey is to move
vmxml.all_channels into the DomainDescriptor.
This allow us to have both nicer code in vm.Vm
and to save us one unneeded XML parsing.

Change-Id: I870cce435b2e26ff33cbd52f2421c1f3ce661707
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34415
Reviewed-by: Dan Kenigsberg <danken@redhat.com>