domain_descriptor: Eliminate magic hash init value
We used to initialize devices hash to 0 when there are no devices. There
is no need for using this magic value. Hashing an empty string does not
require additional comments and is more general.

The magic "0" hash value was exposed in the past in the stats xml, and
could be used for detecting a fresh domain without any devices, but this
value is not exposed any more since we introduced the guest agent disk
mapping hash.

Change-Id: I36767eafbef3dbebef5a1d04417f18e0cc4898c2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34622
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>