guestagent: Simpler and safer disk mapping hash
Previously diskMapingHash was initialized to the magic value 0, or to
hash of a sorted json dump of the diskMapping dict. Now diskMappingHash
is always a real hash, and is read only.

The hash is computed automatically when the value of guestDiskMapping is
changed, and there is no way to forget to update the hash.

Change-Id: Ia914cf9860cef019b3040a3f034e479e6a93ed97
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34625
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>