netlink: event monitor
Current event monitor is implemented as a wrapper around `ip monitor`
command.  It sometimes returns before the underlying monitor is properly
started. However, it's used just in a few tests.

We want to implement more natural version of monitor - wrapper around
the libnl library. With this monitor we will be able e.g. to monitor
`ifup` operation on a new device and guarantee properly created network
before returned.

Change-Id: I0f4fcfde87ad51eb832f54862371b4da1281826e
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/36197
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>