spec: remove redundant 'Requires(X): policycoreutils-python'.
Having an rpm as both 'Required' and 'Required(post/preun) is redundant';
a 'Required' dependency is present for post/preun.

There where two possible solutions here: remove the 'Required(X)'
or remove the 'Required' lines. I choose the first since an rpm
dependency for stage x, is a dependency for all stages before x.
This is done so that rpm will never ask to install a package when
trying to remove one[1]. This means that we practically need
policycoreutils-python for all stages (except postun).

See also:
[1] Context Marked Dependencies:
http://www.rpm.org/max-rpm-snapshot/s1-rpm-depend-manual-dependencies.html

[2] related vdsm patches:
8409, 10000

Change-Id: I293db914e41dde0020cf4da5b5f0b0723c3fac89
Signed-off-by: Mooli Tayer <mtayer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34730
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>