Fix ballooning behaviour of freshly received VM
There is an issue with freshly received VMs. MOM might ask for
the current balloon status before the stats threads collect all
data. The target will be reported as 0 in that case and the VM
will be requested to return all its free memory.

The balloon info method won't return any data before it
has everything ready after this fix is applied.

This also fixes the unittest that was using balloon for checking
whether the stats system works.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1123274
Change-Id: If44c8e9d65b7221b2dec5fdd1f2d7d4343f7aeaa
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/34887
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>