multipath: add specific fail entry for emc-vnx
As few other storage arrays also the emc-vnx is forcing the
queue_if_no_path option on the multipath device.

This patch adds a specific section for the device maintaining
the emc defaults but forcing to fail the io instead of queuing.

In case of queuing we observed that the processes were blocked
on io and among other things it prevented the storage pool to
be reconstructed.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1098769
Change-Id: Icf5f4c89eac1273bf3a8e1afa300fec7b5b68eae
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35072
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
(cherry picked from commit 9880fdf9989718a097bb05571a6344e28c2f258b)
Reviewed-on: http://gerrit.ovirt.org/35179
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>