virt: cleaner getAllVmStats
the api.Global.getAllVmStats method was implemented reusing
api.Global.getVMList and api.VM.getStats().

This lead to unnecessary work, and a lot of layering traversal
with objects being wrapped in responses, which in turn are
promptly discarded and rewrapped again to obtain the final
result. All of this without any clear benefit.

This patch drops all the layering and use simple and lean
methods to obtain the same result doing much less work.
Unfortunately, total performance gains are expected
to be negligible because this function was already
very low in the profiling data.

Related-To: https://bugzilla.redhat.com/1177634
Change-Id: If1ed3f49c236086c6184973762d9909d7f79bf7e
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35140
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>