dummy_nics: add hotplug=no so udev does not ifup them
In EL6 udev/initscripts execute ifup if an ifcfg file exists for the
just plugged device (which applies often for virtual devices). Due
to how troubling this functionality is they removed it from el7/F20+
but in el6 it can disrupt our configuring with parallel ifups,
overriding of bond slaves config, etc.

Change-Id: If07817d51ee8c71a1bcfe7d45f110b884fd516fd
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35089
Reviewed-by: Dan Kenigsberg <danken@redhat.com>