oopTests: replace 'if None' with try except clause
When asserting if the instance is None at the end of the test,
we already assume that it should be None,
or else the test fails.
This means that the 'if None' check is redundant and
can be moved into a try except clause because in case
of test failure we still want the relevant logs.

Change-Id: I8589ec591aad370cb390c0548ec52ded155bf085
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35243
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>