utils: add monotonic_time function
We currently use time.time() to get timestamp
with (near) second precision.
We we just need the difference between two instants
in time, os.times() provide safer alternative
because it is more resilient with respect to
system clock adjustments, and can never go
backwards.

This function add a simple monotonic_time() function
based on os.times().

The drawback of this function is lower resolution.
The actual resolution depend on the kernel configuration,
but a commonly found value is ten milliseconds.

Change-Id: I65c7d6041ed89fc8c497115b2d70341e82936b26
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35350
Reviewed-by: Dan Kenigsberg <danken@redhat.com>