netinfoTests: suggest to refresh bonding defaults in a unit test
With the unfortunate, but pragmatic move from determining the defaults
at runtime to distributing them as a file we might not always be
up to date when the kernel introduces new bonding options.

The unit test will now suggest to run `vdsm-tool dump-bonding-defaults`
when new options are found instead of just failing.

As we moved from reading the defaults from a reference bond we can no
longer trigger read errors, thus assertNotRaises was removed from the test.

Change-Id: I5fad615479fb7edd29235161b4c31a19019389de
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35289
Reviewed-by: Dan Kenigsberg <danken@redhat.com>