jsonrpc: have simplejson as soft dependency
In commit ec58bca5cb18 we added dependency to
the external simplejson package to gain some speed,
especially on RHEL/Centos 6.x.

We discovered the hard way that python-simplejson
was moved out of the base system in RHEL/CentOS 7,
and put into the EPEL repository.
This is alone makes ec58bca5cb18 not good enough.

We still want to get the JSON codec speedups, so
to this patch
* moves the requires to RHEL/Centos 6.x only
* makes simplejson a soft dependency.

Change-Id: I11c05cff5b1cc3727653fa4ce210a76fcadf5628
Bug-Url: https://bugzilla.redhat.com/1155146
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35354
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35405
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>