irs: Directly register domain state change callbacks
Currently, domain state change callbacks aren't actually registered
until a storage pool is connected.  In a hosted engine environment there
is not a traditional storage pool associated with the storage domain
that is hosting the engine VM so in this case the event doesn't get
registered.  Since storage pools are becoming deprecated, there is no
reason to delay callback registration based on connectStoragePool.

Change-Id: I28d4f1af4f94da1a8e3511abbe160ea66e81daad
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1157421
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35369
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>