Instruct MOM to ignore ballooning when guest agent is not running
This makes use of the new MOM feature (valid function) and
associated guest memory collector (GuestMemoryOptional).

The ballooning part of the policy will be ignored when guest agent
is not running while the rest of the policy is still evaluated.

The most notable change is that MOM used to ignore the whole
policy when guest agent was not available.

Change-Id: I20f636cd9290340022d42210cbb5ba6c3b009dd9
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35407
Reviewed-by: Doron Fediuck <dfediuck@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>