clientIF: Use a weakref.proxy when registering contEIOVMs
Our storage Event class is taking a weakref on the callback function
that is passed into Event.register.  Unfortunately weakrefs on bound
methods are dead on arrival.  See this [1] discussion for more
information.  Use the same approach as StoragePool._upgradeCallback and
use a weakref.proxy to build the callback function.

[1] http://code.activestate.com/recipes/81253/

Change-Id: Ib06967b26fc9ace41e5f8305a8fadafbbb0fc4e4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1157421
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35436
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>