lvm: Disable chown hack for EL6
Testing EL6 revealed that the new vdsm lvm udev rules required for EL7
introduced a race which causes createVolume to fail randomly on some
machines.

Since commit 9240b8c5782 (lvm: Do not use udev to set permissions on
vdsm images), devices are created with default permissions - root:disk.
Then udev run chown which set the permissions to vdsm:qemu.  Because of
udev "watch" feature, this permission dance happens twice when creating
a logical volume:

    root:disk -> vdsm:qemu -> root:disk -> vdsm:qemu

Typically when lvcreate returns the permissions changes are done, but
sometimes it may return while the events are processed. Then when we run
qemu-img, it fail because the disk permissions are not vdsm:qemu.

On EL6 we don't need to use the chown hack; We used it because it looked
like a simple solution that works everywhere. So now we use the old udev
rules format on EL6, and the new format (chown hack) on EL7 and Fedora.

By default, we build the old udev format. To enable the new format,
configure with --enable-chown-hack. The spec enables the chown hack
on EL7 and Fedora trough configure.

To create a different rule depending on the configuration, the udev rule
is generated now by a python script.

On EL7 using our chown hack, the race does not exist. We also verified
that using a script creating 600 disks, and using automated QE tests.

Change-Id: I7d84c986511d9cfc7695a79ef5944edf70050934
Bug-Url: https://bugzilla.redhat.com/1160620
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35440
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35486
Reviewed-by: Allon Mureinik <amureini@redhat.com>