drv_name: Make drv_name support unicode types
Our json serialization and de-serialization for both the hooks and
persistence add unicode text to the flows of our codebase.

Before we would only have "foo". Now we have as well u"foo".
In order to make it work in unicode we also need to provide for real
unicode characters u"トトロ".

Change-Id: Ic705eaaad22fb1ed56bca3363a3559ee1b13d63f
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35471
Reviewed-by: Dan Kenigsberg <danken@redhat.com>