virt: make vmdevices a subpackage
This patch transforms the vmdevices module into a subpackage,
to make it suitable to host the remaining, big devices
in nicely separate modules.

A few decisions where made for the sake of practicality
and to make this patch minimal, although meaningful:

- the subpackage retains the 'vm' prefix instead the simpler
  'devices', to reduce name clashes.
- not-empty __init__, hosting subpackage imports and constants.
  Nothing else will go there.
- 'small' devices moved into a generic module named 'core',
  which is the shortest yet meaningful name I could find.

No changes besides import fixes and code being scatthered around.

Change-Id: I66ea868df27b7dd86ba884804016b34721ff7f48
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35482
Reviewed-by: Dan Kenigsberg <danken@redhat.com>