Live Merge: Proactively extend merge target volume
Since libvirt is not yet providing the write watermark information for
all volumes, we need to employ a workaround to support live merge on
block-based storage.  When starting a merge, we can request the base
volume (merge target) to be extended to the size of the top volume
(merge source) plus one chunk to accomodate active writing.  This may
cause some over-extension of the merge target volume but the upside is
that the merge should complete without active monitoring.  Future
versions of oVirt will take advantage of an upcoming libvirt enhancement
to report write watermarks for all volumes which will allow on-demand
volume extension.

Change-Id: I3d483f9dfe6e19a9f9fec84950bf2ea0c7ed7946
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1155583
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35614
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>