networkTests: Make DEFROUTE controllable so the system's gateway is preserved
testSetupNetworksAddDelDhcp, a test on a veth pair (with dnsmasq and
a VDSM-invoked dhclient), has recently been adapted to account for source
routing. Other two tests using dhclient directly however result in
the replacement of the system's default route.

Fix this situation on Fedora and RHEL by instructing their dhclient-script
not to set the gateway on the veth.

Change-Id: I0c0210b188feec45d151060d5cadaefe1ed6dc99
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/35701
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>